nlp project/requirements.txt
streamlit==1.32.0
spacy==3.7.4
nltk==3.8.1
pandas==2.2.1
openai==1.12.0
python-dotenv==1.0.1 